Implment global navigation handling
Based on code provided here:
http://devcenter.wintellect.com/jprosise/handling-the-back-button-in-windows-10-uwp-apps
