+++ MasterDetail-Native/MasterDetail.UWP/Views/AddItems.xaml
@@ -37,7 +37,7 @@
 
     <Page.BottomAppBar>
         <CommandBar>
-            <AppBarButton Label="Save" Icon="Save" Click="btnAddItem_Click"/>
+            <AppBarButton Label="Save" Icon="Save" Click="SaveItem_Click"/>
         </CommandBar>
     </Page.BottomAppBar>
 </Page>
