+++ MasterDetail-Native/MasterDetail.UWP/Views/BrowseItemDetail.xaml.cs
@@ -1,22 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
+﻿using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 using MasterDetail.Model;
 using MasterDetail.ViewModel;
-using Windows.UI.Core;
-
-// The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=234238
 
 namespace MasterDetail.UWP.Views
 {
@@ -32,24 +18,8 @@ namespace MasterDetail.UWP.Views
         {
             this.InitializeComponent();
 
-            
-            SystemNavigationManager.GetForCurrentView().BackRequested += OnBackRequested;
-
             ViewModel = new ItemDetailViewModel();
             DataContext = new ItemDetailViewModel();
-
-
-        }
-
-        private void OnBackRequested(object sender, BackRequestedEventArgs e)
-        {
-            Frame rootFrame = Window.Current.Content as Frame;
-
-            if (rootFrame.CanGoBack)
-            {
-                e.Handled = true;
-                rootFrame.GoBack();
-            }
         }
 
         public void Show_Click(object sender, RoutedEventArgs e)
@@ -58,12 +28,10 @@ namespace MasterDetail.UWP.Views
         }
         protected override void OnNavigatedTo(NavigationEventArgs e)
         {
-
             ViewModel.Item = (Item)e.Parameter;
 
             txtText.Text = ViewModel.Item.Text;
             txtDesc.Text = ViewModel.Item.Description;
-
         }
     }
